refactor: add index_spec for discounts instead of using merchant dashboard to test discounts index view
